save last changes

+++ app/src/main/java/com/mif50/instagramimagefilter/Interface/EditImageFilterListener.kt
@@ -2,8 +2,8 @@ package com.mif50.instagramimagefilter.Interface
 
 interface EditImageFilterListener {
     fun onBrightnessChanged(brightness: Int)
-    fun onSaturationChanged(saturation: Int)
-    fun onConstrainChanged(constraint: Int)
+    fun onSaturationChanged(saturation: Float)
+    fun onConstrainChanged(constraint: Float)
     fun onEditStarted()
     fun onEditCompleted()
 }
